[ISSUE-6] Modularized LoginForm and OAuth buttons
