+++ public/locales/en/login.json
@@ -1,10 +1,3 @@
 {
-    "loginTitle": "Login",
-    "emailInputError": "Please enter an email",
-    "emailInputPlaceholder": "Email",
-    "passwordInputError": "Please enter a password",
-    "passwordInputPlaceholder": "Password",
-    "loginButton": "Login",
-    "noAccount": "Don't have an account? <1>Register now!</1>",
-    "oauthSubdivision": "Or login via"
+    "loggingIn": "Logging in..."
 }
