+++ public/locales/en/loginForm.json
@@ -1,0 +1,10 @@
+{
+    "loginTitle": "Login",
+    "emailInputError": "Please enter an email",
+    "emailInputPlaceholder": "Email",
+    "passwordInputError": "Please enter a password",
+    "passwordInputPlaceholder": "Password",
+    "loginButton": "Login",
+    "noAccount": "Don't have an account? <1>Register now!</1>",
+    "oauthSubdivision": "Or login via"
+}
